node clustering implemented
